requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
rank-bm25==0.2.2
# ragatouille - DISABLED (replaced with sentence-transformers for dense retrieval)
qdrant-client==1.7.0
neo4j==5.14.0
redis==5.0.1
google-generativeai==0.3.2
langchain==0.1.0
sentence-transformers==2.2.2
numpy>=1.24.0
scikit-learn>=1.3.0
spacy==3.7.2
nltk==3.8.1
pydantic==2.7.4
python-dotenv==1.0.0
loguru==0.7.2
celery==5.3.4
prometheus-client==0.19.0
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-instrumentation-fastapi==0.42b0
PyPDF2==3.0.1
python-multipart==0.0.6
python-docx==0.8.11
